backend/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
Flask==3.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
packaging==24.0
pycparser==2.22
pycryptodome==3.20.0
pymongo==4.7.2
python-dotenv==1.0.1
Werkzeug==3.0.3
